# [REVIEW] DataFrame `insert` and creation optimizations