This PR removes double index equality & reindexing checks in `DataFrame` construction and `insert` code-flows.

`scalar_broadcast_to`, a **2.2x speedup** for `str` dtype, numeric types perf remains the same:
```python

# This PR:

In [12]: %timeit cudf.utils.utils.scalar_broadcast_to('abc', 400000000, "str")
100 ms ± 208 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# branch-22.04
In [25]: %timeit cudf.utils.utils.scalar_broadcast_to('abc', 400000000, "str")
227 ms ± 10 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```



`assign`, a **1.2x-1.4x speedup**:
```python

In [9]: import numpy as np

In [10]: df = cudf.DataFrame(
    ...:     {
    ...:         "a": [1, 2, 3, np.nan] * 10000000,
    ...:         "b": ["a", "b", "c", "d"] * 10000000,
    ...:         "c": [0.0, 0.12, np.nan, 10.12] * 10000000,
    ...:     },
    ...: )

# THIS PR
In [11]: %timeit df.assign(f=10)
28.1 ms ± 365 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# branch-22.04
In [4]: %timeit df.assign(f=10)
35 ms ± 10.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)



# THIS PR
In [4]: %timeit df.assign(g='hello world')
38.4 ms ± 250 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# branch-22.04
In [4]: %timeit df.assign(g='hello world')
54 ms ± 82.6 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```





DataFrame constructor, a **2x speedup**:

```python

In [2]: s1 = cudf.Series([1, 2, 3] * 1000000)

In [4]: s1.index = s1.index.astype("float64")

In [5]: s2 = cudf.Series([2, 3, 3] * 1000000)

In [6]: s2.index = s2.index.astype("float64")


In [12]: s3 = cudf.Series([10, 11, 12] * 1000000)

In [13]: s3.index = s3.index.astype("float64")

# THIS PR
In [14]: %timeit cudf.DataFrame({'a':s1, 'b':s2, 'c':s3})
2.81 ms ± 63.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# branch-22.04
In [9]: %timeit cudf.DataFrame({'a':s1, 'b':s2, 'c':s3})
5.37 ms ± 188 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
<!--

Thank you for contributing to cuDF :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. There are CI checks in place to enforce that committed code follows our style
   and syntax standards. Please see our contribution guide in `CONTRIBUTING.MD`
   in the project root for more information about the checks we perform and how
   you can run them locally.

4. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

5. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please mark your pull request as Draft.
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#converting-a-pull-request-to-a-draft

6. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove it from "Draft" and make it "Ready for Review".
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#marking-a-pull-request-as-ready-for-review

   If assistance is required to complete the functionality, for example when the
   C/C++ code of a feature is complete but Python bindings are still required,
   then add the label `help wanted` so that others can triage and assist.
   The additional changes then can be implemented on top of the same PR.
   If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

7. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on the target branch, force push, or rewrite history.
   Doing any of these causes the context of any comments made by reviewers to be lost.
   If conflicts occur against the target branch they should be resolved by
   merging the target branch into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->